Smarter slime-parse-form-upto-point
I note that the current implementation of this function sometimes goes
beyond point (uses forward-sexp).  The name is a lie.

With this commit, when point is right after an opening parenthesis, try
to include whatever symbol comes next rather than just adding an empty
string before the cursor marker.  Then autodoc can show the appropriate
arglist documentation for the form.