swank-repl: a repl-specific dispatch table variable
This makes it possible to do something like:

  (setf swank-repl:*repl-pprint-dispatch*
        (copy-pprint-dispatch))

  (set-pprint-dispatch 'integer
                       (lambda (stream object)
                         (write-char #\# stream)
                         (write-char #\x stream)
                         (write object :stream stream :base 16 :pretty nil))
                       0
                       swank-repl:*repl-pprint-dispatch*)

Now integers will be printed in hexadecimal notation (with #x
indicator) in the REPL, while SLIME (and other code) can happily
continue to pretty-print assuming a more traditional pprint dispatch
table.  The assumption may not be warranted, but this is the real
world.